% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm25.R
\docType{data}
\name{pm25}
\alias{pm25}
\title{Tibble con i dati di pm25 per 14 stazioni della regione PUGLIA}
\format{
Un tibble con 8 colonne e 20194 osservazioni
}
\usage{
pm25
}
\description{
Tibble con i dati di pm25 per 14 stazioni della regione PUGLIA
}
\keyword{datasets}
